\name{entryData-methods}
\docType{methods}
\alias{entryData,movG-method}
\title{ Methods for Function entryData }
\description{
  Methods for function \code{entryData} 
}
\section{Methods}{
\describe{

  \item{object = "movG"}{returned is a \code{data.frame} with:
    row, column, adj. phenotypic value,
    obs. phenotypic value, moving mean, number of values}
  
}}
\keyword{methods}
