\name{fitted-methods}
\docType{methods}
\alias{fitted,movG-method}
\title{Methods for Function fitted in Package "stats"}
\description{
  Methods for function \code{fitted} in Package "stats" 
}
\section{Methods}{
\describe{

\item{object = "ANY"}{default method, the S3 generic}

\item{object = "movG"}{Returns a vector with the adjusted values.}
}}
\keyword{methods}

