\name{movingMean-methods}
\docType{methods}
\alias{movingMean,movG-method}
\title{ Methods for Function movingMean }
\description{
  Methods for function \code{movingMean} 
}
\section{Methods}{
\describe{

  \item{object = "movG"}{returns a vector with the
    moving means}
  
}}
\keyword{methods}
