\name{residuals-methods}
\docType{methods}
\alias{residuals,movG-method}
\title{Methods for Function residuals in Package `stats'}
\description{
  Methods for function \code{residuals} in Package `stats' 
}
\section{Methods}{
\describe{

\item{object = "ANY"}{default method, the S3 generic}

\item{object = "movG"}{Residuals from the model to calculate the
  regression coefficient for adjustment by the covariate}
}}
\keyword{methods}

