\name{show-methods}
\docType{methods}
\alias{show,movG-method}
\title{Methods for Function show in Package `methods'}
\description{
  Methods for function \code{show} in Package `methods'
}
\section{Methods}{
\describe{

\item{object = "movG"}{Some summary statistics and informations are
  calculated from the object and printed. }
}}
\keyword{methods}

