\name{sketchGrid-methods}
\docType{methods}
\alias{sketchGrid,ANY-method}
\title{Methods for Function sketchGrid in Package "mvngGrAd"}
\description{
  Methods for function \code{sketchGrid} in Package "mvngGrAd"
}
\section{Methods}{
\describe{\item{objects:}{i = "ANY",
    j = "ANY", rowLimit = "ANY",colLimit = "ANY",
    
    layers = "ANY", shapeCross = "ANY",
    
    excludeCenter = "ANY"

    Method to sketch designed grid, for example for usage with function
    \code{movingGrid}.   
  }

  
}
}

\keyword{methods}

